backend/core-service-ms/requirements.txt
asgiref==3.9.1
Django==4.2.24
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
redis==5.0.1
django-redis==5.4.0
